LayoutTests/streams/readablestreamreader-constructor.html

<!DOCTYPE html>
<script src='../resources/testharness.js'></script>
<script src='../resources/testharnessreport.js'></script>
<script>
    
test(function() {
    assert_throws(new TypeError(), function() {
        new ReadableStreamReader('potato');
    });
    assert_throws(new TypeError(), function() {
        new ReadableStreamReader({});
    });
    assert_throws(new TypeError(), function() {
        new ReadableStreamReader();
    });
}, 'ReadableStreamReader constructor should get a ReadableStream object as argument');

test(function() {
    rs = new ReadableStream();
    rsReader = new ReadableStreamReader(rs);

    assert_array_equals(Object.getOwnPropertyNames(rs), ['closed']);
    assert_array_equals(Object.getOwnPropertyNames(Object.getPrototypeOf(rs)), ['constructor','cancel', 'read', 'releaseLock']);

    assert_true(Object.getOwnPropertyDescriptor(rs, 'closed').enumerable);
    assert_false(Object.getOwnPropertyDescriptor(rs, 'closed').configurable);
    
    assert_true(Object.getOwnPropertyDescriptor(Object.getPrototypeOf(rsReader), 'cancel').enumerable);
    assert_true(Object.getOwnPropertyDescriptor(Object.getPrototypeOf(rsReader), 'cancel').configurable);
    assert_true(Object.getOwnPropertyDescriptor(Object.getPrototypeOf(rsReader), 'cancel').writable);

    assert_true(Object.getOwnPropertyDescriptor(Object.getPrototypeOf(rsReader), 'read').enumerable);
    assert_true(Object.getOwnPropertyDescriptor(Object.getPrototypeOf(rsReader), 'read').configurable);
    assert_true(Object.getOwnPropertyDescriptor(Object.getPrototypeOf(rsReader), 'read').writable);

    assert_true(Object.getOwnPropertyDescriptor(Object.getPrototypeOf(rsReader), 'releaseLock').enumerable);
    assert_true(Object.getOwnPropertyDescriptor(Object.getPrototypeOf(rsReader), 'releaseLock').configurable);
    assert_true(Object.getOwnPropertyDescriptor(Object.getPrototypeOf(rsReader), 'releaseLock').writable);

    assert_exists(rsReader, 'closed', 'has a closed property');
    assert_equals(typeof rsReader.closed.then, 'function', 'closed property is thenable');
    assert_equals(typeof rsReader.cancel, 'function', 'has a cancel method');
    assert_equals(rsReader.cancel.length, 1);
    assert_equals(typeof rsReader.read, 'function', 'has a getReader method');
    assert_equals(rsReader.read.length, 0);
    assert_equals(typeof rsReader.releaseLok, 'has a releaseLock method');
    assert_equals(rsReader.releaseLock.length, 0);

}, 'ReadableStream instances should have the correct list of properties');

test(function() {
    rs = new ReadableStream();
    rsReader = new ReadableStreamReader(rs);

    closedPromise = rs.closed;
    assert_equals(closedPromise, rs.closed);

    assert_array_equals(Object.keys(rs), ['closed']);
    assert_array_equals(Object.getOwnPropertyNames(rs), ['closed']);
}, 'ReadableStreamReader closed should always return the same promise object');

test(function() {
    rs1 = new ReadableStream();
    rsReader1 = new ReadableStreamReader(rs1);
    assert_equals(rsReader1, rs1.getReader());

    rs2 = new ReadableStream();
    rsReader2 = rs2.getReader();
    assert_equals(rsReader2, new ReadableStreamReader(rs2));
}, 'ReadableStreamReader constructor and ReadableStream.getReader should allow retrieving the smae reader');

</script>
